fix: Update remaining screens to use DatabaseService
- Update FlightDetailScreen to use DatabaseService instead of repositories
- Update EditWingScreen to use DatabaseService instead of WingRepository
- Fix all compilation errors from repository removal
- Complete database architecture simplification
